[New] Added test cases for pages support components navigation
