+++ app/javascript/components/pages/support/components/navigation/component.jsx
@@ -33,7 +33,15 @@ const Component = ({ css, handleToggleNav, menuList, mobileDisplay, onClick, sel
     };
 
     return (
-      <ListItem button key={id} onClick={handleClick} selected={selected} disabled={disabled} classes={classes}>
+      <ListItem
+        data-testid="list-item"
+        button
+        key={id}
+        onClick={handleClick}
+        selected={selected}
+        disabled={disabled}
+        classes={classes}
+      >
         <ListItemText key={id} primary={text} />
         {hasUnsubmittedOfflineChanges && id === SUPPORT_FORMS.resync && <Jewel isForm />}
       </ListItem>
@@ -42,7 +50,9 @@ const Component = ({ css, handleToggleNav, menuList, mobileDisplay, onClick, sel
 
   return (
     <ConditionalWrapper condition={mobileDisplay} wrapper={Drawer} {...drawerProps}>
-      <List component="nav">{renderList}</List>
+      <List data-testid="list" component="nav">
+        {renderList}
+      </List>
     </ConditionalWrapper>
   );
 };
